(refactor(enrich-original-pub-year)): split CLI and enrichment modules
- move CLI parsing into a dedicated cli module and expose helpers lazily from package init
- add a python -m entrypoint and point the VS Code launch config at the new module target
- extract enrichment workflow and matching utilities into focused modules while keeping public API stable
